+++ core/src/main/java/tech/pegasys/web3signer/core/config/BaseConfig.java
@@ -63,5 +63,5 @@ public interface BaseConfig {
 
   boolean keystoreParallelProcessingEnabled();
 
-  int getWorkerPoolSize();
+  int getVertxWorkerPoolSize();
 }